+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [2.0.0] - 2017-06-24
+
+Complete re-write.
+
+### Breaking Changes
+
+* New API
+
+### New Features
+
+* Flat template output mode.
+
+* Saving dictionary files.
+
 ## [1.0.0] - 2016-10-15
 
 Initial release.